+++ connectors/yarpc/helpers.go
@@ -185,6 +185,30 @@ func RPCTypeToClientType(t dosarpc.ElemType) dosa.Type {
 	panic("bad type")
 }
 
+// RPCTagsFromClientTags converts tags to the RPC version.
+func RPCTagsFromClientTags(tags map[string]string) []*dosarpc.FieldTag {
+	rpcTags := make([]*dosarpc.FieldTag, 0, len(tags))
+	for k, v := range tags {
+		name := k
+		value := v
+		ft := dosarpc.FieldTag{Name: &name, Value: &value}
+		rpcTags = append(rpcTags, &ft)
+	}
+	return rpcTags
+}
+
+// RPCTagsToClientTags converts thrift tags to a map.
+func RPCTagsToClientTags(rpcTags []*dosarpc.FieldTag) map[string]string {
+	if len(rpcTags) == 0 {
+		return nil
+	}
+	tags := map[string]string{}
+	for _, t := range rpcTags {
+		tags[*t.Name] = *t.Value
+	}
+	return tags
+}
+
 // PrimaryKeyToThrift converts the dosa primary key to the thrift primary key type
 func PrimaryKeyToThrift(key *dosa.PrimaryKey) *dosarpc.PrimaryKey {
 	ck := make([]*dosarpc.ClusteringKey, len(key.ClusteringKeys))
@@ -234,7 +258,7 @@ func entityDefToThrift(ed *dosa.EntityDefinition) *dosarpc.EntityDefinition {
 	fd := make(map[string]*dosarpc.FieldDesc, len(ed.Columns))
 	for _, column := range ed.Columns {
 		rpcType := RPCTypeFromClientType(column.Type)
-		fd[column.Name] = &dosarpc.FieldDesc{Type: &rpcType}
+		fd[column.Name] = &dosarpc.FieldDesc{Type: &rpcType, Tags: RPCTagsFromClientTags(column.Tags)}
 		cols = append(cols, column.Name)
 	}
 
@@ -287,7 +311,7 @@ func FromThriftToEntityDefinition(ed *dosarpc.EntityDefinition) *dosa.EntityDefi
 		fields[i] = &dosa.ColumnDefinition{
 			Name: colName,
 			Type: RPCTypeToClientType(*v.Type),
-			// TODO Tag
+			Tags: RPCTagsToClientTags(v.Tags),
 		}
 	}